refactor: move wapps
