+++ frontend/src/lib/components/wapps/InstallWAppAmount.svelte
